+++ innopoints/schemas/__init__.py
@@ -1,4 +1,5 @@
 from .project import *
 from .account import *
+from .activity import *
 
 from .schemas import *  # temporary until moved all schemas
